Add an Alpha version of zsvjmp
The trick with this one, which I missed at first, was to set register
27 (pv) to the address of setjmp before branching into it. The alpha
procedure calling sequence assumes that the caller of a procedure does
this and it is required to set the global pointer correctly for the
procedure being executed.

The STScI version of zsvjmp (produced by technical support at DEC)
could not be used as it was just setjmp with the name changed.
The status argument used by zsvjmp was being ignored, and setjmp
was being called with the wrong jump buffer address.