Fix declaration of cdsmem in rpp
This fixes a problem that rpp.e does not work correctly when build
with optimization: since cdsmem is initialized everywhere as `integer
mem(1)` the C compiler (after f2c) assumes that the variable is not an
array, but a pointer to a single variable, and any subsequent write
can be replaced by just writing the last value.

It should really be a `integer mem(60000)`.