Accept zero date in archives
A zero date is a valid (and not anymore unplausible) entry in an
library archive. Archives with that entries can f.e. created with
newer GNU ar and the `D` option, or when the
`--enable-deterministic-archives` option was set during the build of
ar.

To allow at least some updating, we use the last modified date of the library if the date is set to zero.