Adjust lex processing in generic for newer lex (flex & Co.).
The hack to replace the getc() in the input does not work anymore due
to structural changes in the generated code (use of fread(), ferror(),
fileno()) etc.

On the other hand, the custom k_getc() was introduced only because VMS
could not correctly do fseek()/ftell() (this is at least from the
comments in the code). Since VMS is not supported anymore (and the
problem is there probably also fixed), we cann revert back to the
standard solutions.