Fix pointer type for ZCLSND call
The first argument of ZCLSND is an `XINT *` (aka `long int *`, 64bit), while
`vm_server` is `int` (32 bit) only. Just (implicitly) changing pointer type
destroys the alignment and may caise an illegal access of the address after
`*vmserver`. And on big endian, even the wrong value is transferred.

Since the variable is not an output, we can just assing a temporary var with
the correct type. And just adjust the type of the status var.