Fix pointer type for ZAWTBF call
The first argument of ZAWTBF is an `XINT *` (aka `long int *`, 64bit), while
`pltr.status` is `int` (32 bit) only. Just (implicitly) changing pointer type
destroys the alignment and may raise an illegal access. On big endian, even
the wrong value is transferred.

We use the (already defined for that purpose) `status` variable with the
correct type here to get the correct value.