Fix pointer type for BYTMOV call
The last argument of BYTMOV is an `XINT *` (aka `long int *`, 64bit), while
`nbytes` is `int` (32 bit) only. Just (implicitly) changing pointer type
destroys the alignment and may raise an illegal access. On big endian, even
the wrong value is transferred.

Since the variable is not an output, we can just assing a temporary var with
the correct type.