Fix linenum calculation in xpp
First, when `begin` is parsed, the end of line `\n` is not eaten
up. In fact, there may be code following after `begin` on the same
line; this is placed after xpp in the line after `begin`. If nothing
follows the `begin` statement on the same line, an empty line is
added. This means, that `linenum` must __not__ increased here; it is
increased when the `\n` of this line is found.

Similar argument for parsing a `task` line; `linenum` must
__not__increased here as well.

When parsing `.help` ... `.endhelp` however, one `linenum` increase was
missing, which is added now.