Add more tests, taken from the tutorial
http://iraf.noao.edu/iraf/ftp/iraf/misc/exer211.tar.Z

To All Interested Parties:

A set of IRAF exercises is available for the novice user.  The exercises
were developed under IRAF Version 2.10.2 but have been revised for IRAF
Version 2.11. (Please be sure you are using the right version of these
exercises for your IRAF installation.)  The exercises are complete with
datasets.  Exercises 2 and 4 share the same dataset; 2 reduces the data that
is then used in 4. Exercises 3 and 5 also share the same dataset; 3 must be
executed before 5. An image display is necessary for exercises 1 and 4, but
is optional for the others.

   1) intro  - an introduction to the basics of IRAF

>> 2) ccd1   - a look at CCD reductions for direct images <<<<<<<<<<<<<<<<<<

   3) ccd2   - CCD reductions for spectroscopic data

   4) phot   - an introduction to aperture photometry using APPHOT and the
               steps for photometric calibration using PHOTCAL

   5) spec   - the extraction and wavelength calibration of slit spectra from
               2-d CCD images

   6) dataio - an exercise in reading and writing tapes