Remove check for negative pointers (indices) in mfree
Since ptr is relative to the `Mem` common block, it may be also
negative. Therefore, the sign of the pointer cannot be used to determine
an already freed pointer. This is now done by setting the pointer to NULL.

It appears that the pointer was set to NULL anyway already for some unknown
reason, however.