Handle negative pointer in garbage collector
Previously, the garbage collection in `mgc.x` used two pseudo values for
organizational issues: implicitly initialized NULL for the end of the `MGC`
buffer, and negative values for pointers that are already freed.

Since pointers may be negative (they are relative to the `Mem` common block),
this patch changes this behaviour: we now use NULL for pointers that are
already freed, and an additional variable `bmax` in the `/nmemio/` common
block to determine current the end of the `MGC` buffer.

Note that this could not be fully tested, since the garbage collector
is active only when the environment variable `MEMIO_COLLECT` is set to
1. This mode is however already buggy with the precompiled version
2.16.1:

```
$ MEMIO_COLLECT=1 cl
setting terminal type to 'xgterm' ...

PANIC in `/iraf/iraf/bin/64/bin.linux64/x_system.e': invalid SET in IRAF Main
```