Handle negative pointer in garbage collector
Previously, the garbage collection in `mgc.x` used two pseudo values for
organizational issues: implicitly initialized NULL for the end of the `MGC`
buffer, and negative values for pointers that are already freed.

Since pointers may be negative (they are relative to the `Mem` common block),
this patch changes this behaviour: we now use NULL for pointers that are
already freed, and an additional variable `bmax` in the `/nmemio/` common
block to determine current the end of the `MGC` buffer.