Add a trailing `\n` to the end of variable format strings.
f2c does not recognize the length of a variable FORMAT string and will
try to read the string until a non-blank appears.

The format strings are declared as `CHARACTER*8` and do not
automatically have a `\0` appended at the end. Therefore, when the
string is filled with blanks until the end, it will read one char over
the end (which is hopefully a non-blank).

This is probably a bug in f2c; however it is not fixed in the recent
versions as well. To work around this, a trailing `\0` is
automatically added.