Remove vtel subpackage from imred
vtel was used for the data reduction of the Kitt Peak Vacuum
Telescope.  After 29 years of solar imaging, the vacuum telescope was
decommissioned in 2002, which is 15 years ago.

The vtel subpackage uses a the "MRQMIN" routine which was copied from
the Numerical Recipes book. This code is not licensed for public
redistribution and must be removed from the public IRAF package. Since
the subpackage is specialized for data from an instrument that is out
of operation for a long time, it is not worth replacing the NR code
with free code.

Therefore, this patch removed the vtel package completely.