Use str*n*cpy and s*n*printf to fill file header.
This is required at least for the size and mtime fields (which are
both defined as char[12]: sprintf writes 12 chars, plus the closing '\0',
which would give 13 chars into fields with the length of 12. To avoid
this, we limit the max number of chars to the field size; the '\0' is
not needed here anyway.