Check fp for NULL in envinit before trying to close.
When a file to be opened could not be found, the outer loop in the function is
left via `break`, but then the file pointer `fp` is tried to be closed, which
leads to a segementation fault (since it is `NULL`).

This patch checks the file pointer to be not `NULL` before closing and fixes
this.