Fix allocated size of Dimblock
Generally, the size of a structure is not just the sum of the sizes of
its members, since padding may require additional space. For
`Dimblock`, this is the case, since the first member is an `int`, and
the next one is a pointer: on 64-bit platforms, `int` is 4 bytes,
while pointers are 8 bytes and therefore 4 bytes padding are inserted.

This patch fixes the allocated size by using the real size of the
structure.

Also, the length of the `dims` member is set to unspecified, to avoid
compilers to optimize out the access to the member.