photcal: Fix access to fit parameter table
The parameter table contains all fit parameters. These are changed by
the fit subroutine (NLFIT), which expects to have them in a
homogenious "real" (4 byte) array, with subsequent entries directly
neighored. Therefore, the array cannot be set with `Memr[P2R(base+index)]`,
but needs to be set with `Memr[P2R(base)+index]`. This patch changes
this in a consistent way.