Replace "OpenDetachedProcess" with "ForkDetachedProcess" in CL
For background operation, CL needs to start an identical copy of
itself. This was (unnecessarily) emulated by starting a new CL process
and restoring its internal state later.

On POSIX systems, this is not necessary, since the fork() system call
already provides an identical copy of the process.

To utilize the IRAF job management, the fork call needs to be wrapped
in several levels:

- unix/os/zopdr.c provides a SPP callable C interface of fork()
- sys/etc/prfodpr.x is the SPP implementation (incl. job management)
- sys/libc/cprdet.c provides a C interface to the SPP function
- pkg/*cl/bkg.c finally calls that

And one has to provide the naming transition as well.