Use O_RDONLY & Co. in open() calls
From the man page of open (2):

> ```
> int open(const char *pathname, int flags);
> ```
> [...]
>
> The argument flags must  include  one  of  the  following  access modes:
> `O_RDONLY`, `O_WRONLY`, or `O_RDWR`.  These request opening the file
> read-only, write-only, or read/write, respectively.

Some places in IRAF however use `0` for the flags when the file is
opened readonly. This fails in the moment when `O_RDONLY` is not `0`,
as on GNU HURD. Similar of `O_WRONLY` and `O_RDWR`.

This patch uses the pre-defined modes in these cases.