Update wait status a background process before querying it
The original code just signalled the process with signal=0 to check
whether the process still exists. However, when a process is
terminated but still no wait() was called, the process is still in the
"zombie" state, which lets the kill(pid, 0) succeed.

To overcome this, wait() needs to be called before. This is however
the only possibility to retrieve the exit status (subsequent wait()
calls to the same pid will return a failure), so the exit status needs
to be stored in prwait.c, so that it can be retrieved later with
prwait().