Add fork() infrastructure to the IRAF system libs
For background operation, CL needs to start an identical copy of
itself. This was (unnecessarily) emulated by starting a new CL process
and restoring its internal state later.

On POSIX systems, this is not necessary, since the fork() system call
already provides an identical copy of the process.

To utilize the IRAF job management, the fork call needs to be wrapped
in several levels:

- unix/os/zopdr.c provides ZFODPR(), a SPP callable C interface of fork()
- sys/etc/prfodpr.x has PRFODPR, the SPP implementation (incl. job management)
- sys/libc/cprdet.c provides c_prfodpr(), a C interface to the SPP function
- pkg/*cl/bkg.c will finally call that.

And one has to provide the naming transition as well.