Remove file permissions check from install script
This was a convenience function that is not really needed: usual
operations (untar, git checkout etc., moving) don't change the
permission. In the last 30 years, the users should have learned that
they need to keep the permission of files and therefore use "-a" or so
when copying.

On the other hand, the check fails on source-only trees leaving a
irritating "failed" message.