Replace readline library by libedit on MacOS X
Libedit is available as a standard package on MacOSX:

https://opensource.apple.com/source/libedit/

as well as on all Linux distributions. Its license is BSD-3-Clause.
Another option is the "editline" library which is available on Linux
(seems to be outdated, but may be a good replacement for static
linking due to its size).

All three are compatible and can be replaced by changing library in
the linking step of ecl:

 * readline "-lreadline"
 * libedit  "-ledit"
 * editline "-leditline"

References:

* readline https://tiswww.cwru.edu/php/chet/readline/rltop.html
* libedit  http://thrysoee.dk/editline/
* editline https://github.com/troglobit/editline

Since readline is standard on Linux, and there is also a compatibility package
for readline5, we use editline only on MacOS X; however we use the system
provided readline on Linux and therefore can remove its sources from IRAF.

We also clean up the mkpkg file a bit and remove SPARC and STScI specific
stuff.