Remove all non-online documentation from doc/
All this documentation is hopelessly outdated, in ancient formats (groff) or
does not build anymore (LaTeX). In this form, it is useless for the end user.

We might think of re-adding f.e. the site managers guide or the CL manuals at
some point; however only in an up-to-date form. Otherwise, the PDF
documentation (which unfortunately often comes without source) should be
used.