Mention Chisato Yamauchi as copyright owner of the amd64 zsvjmp.s code
The iraf64 web page states for their own code (which is there at least since
2008) https://www.ir.isas.jaxa.jp/~cyamauch/iraf64/index.html

Updates for x86_64
    Wrote an assembler code for Linux x86_64 architecture.

The code here is identical, except for white spaces. I don't think this is a
random coincidence, since the code is in both cases preceded with the
identical line

which only appears on the x86_64 assembler files. So, I am sure that the IRAF
code for amd64 zsvjmp was "taken over" from the iraf64 project.