Fix buffer length in urlget.x
Sometimes, the HTTP response header is larger than SZ_PATHNAME (512
chars). This may happen especially when the URL redirects to a content
delivery service. In this case, the end of the redirected URL was
cut (resulting in a failing download), or the program even segfaulted.

This patch increases the size of the reply buffer to SZ_BUF (8192
chars), and also takes care that the program does not crash when the
header is larger. The response URL however is still limited to
SZ_LINE (1023 chars).

Note that these limits are still arbitrary and may be too short: the
maximum *recommended* size for an URL is 2048, and the http response
header has no size limit at all.  SZ_LINE is however builtin in many
places of string processing, so it is probably not easy to increase
the max URL size to the recommended one (except increasing SZ_LINE
itself).