Remove the creation of the `/iraf` link in system installation
The `/iraf` link contradicts any file system standards and is buggy by
design: It silently requires that the IRAF directory is called `iraf`
by itself, which is not always the case. This symlink is also not
necessary at all, since the proper install path is determined with the
link of `/usr/include/iraf.h` or the `iraf` environment variable.