Rename mkfloat.sh to mkfloat
mkfloat is a script that is called when external packages are
built. The specific implementation of the script (which interpreter)
is not relevant, so the script should not have a interpreter specific
suffix.

This PR removes the suffix in the file name and in all local
invocations. Additionally, external packages need to replace
`mkfloat.csh` and `mkfloat.sh` by `mkfloat`.