Fix cpu_time calculation in unix/os/zgtenv.c
This should return the CPU time in milliseconds. In early versions,
this was calculated as floating point cpu*1000./clkfreq with clkfreq
being the clock ticks per second. At some point, floating point was
tried to be avoided in the kernel, and so this got an (all integer)

    cpu * 1000 / clkfreq

To prevent (undiscovered) overflow for large values of cpu (namely
when cpu > MAX_LONG/1000), the order is reversed:

     cpu / clkfreq * 1000

Note that this is still the same operation, just a different order.

In version 2.15, determination of clkfreq seemed to be complicated,
and this got the buggy

    if (cpu > MAX_LONG/1000)
        *cpu_time = cpu / 10;
    else
        *cpu_time = cpu * 10;

which is wrong, because is assumes clkfreq=100, and it also does the
wrong thing for cpu > MAX_LONG/1000.

In POSIX, the clock frequency can be determined with
sysconf(_SC_CLK_TCK), which is used here now.