Use "macos64" for non-intel 64 bit macOS architectures
For Intel, we still keep the old name "macintel" for backward
compatibility; but for ARM64 (and possibly for other 64-bit macOS
architectures) we use a common name.

This is similar to the philosophy in Linux and FreeBSD: Encode only OS
name and 32/64 bit, but not the CPU architecture.

On this occasion, also remove some architectures which are really outdated