Use the IRAF epoch as the base for zgmtco calculation
This reverts commit ef3151200c8422745a0d2c606e0260f05f036379.
and commit d23c020ccf4b2d2f47e3de376fb4407986d329e0.

and a few others. It basically reverts to the version 2.11.

zgmtco gives the time difference for the following numbers:

 * Seconds since 1980-01-01 local time ("LST")

 * Seconds since 1980-01-01 UTC ("GMT")

Both are (and should be) monotone, so the difference is just the
start. Therefore, we calculate the difference on the base of the begin
of the epoch.

Adding the current (!) summer time (as done in version 2.12) seems to
be a wrong idea here.