Force using POSIX shell in extpkg.cl script
Some shells behave differently; namely zsh prints a warning itself
when a glob doesn't match.  This is f.e. the case when one tries to
get the list of external packages from an empty dir with

    /bin/ls -1ad [a-y]* 2> /dev/null

With zsh, this leads to a message "no matches found" which disturbs
f.e. the tests.

The solution is to use git add unix/hlib/extpkg.cl in
the cl script so that /bin/sh is forced to be used.