Avoid using ECHO in generic's token parser
ECHO does an unconditional output. When processing a condition with f.e.

    $for (rd)
    $if (datatype == r)
        if (IS_INDEFD(GM_XO(fit)))
           call gsset (sf1, GSXREF, INDEFR)
    $endif
    $endfor

then we want the output for IS_INDEFD and INDEFR only while processing
"r" (real). When INDEF(R|D|...) is causing an ECHO, this results in
additional "INDEFDINDEFR" in the output.

Replacing ECHO with outstr(yytext) solves the problem.