Fix keyword search in envinit
The searchs works for "set" or "reset" and is done *after* stepping
over possible whitespaces. The bug here was that "set" resp. "reset"
were searched in the string *before* whitespace removal, but the
result was applied to the string *after* the removal. This works for

set foo = bar

but not for ("_" = space)

___set foo = bar

It is funny to see that this bug is there since 35 years, and nobody
discovered or fixed it...

For us this is important, since otherwise the "noao" variable was not
read in properly.