Replace SIGFUNC by sighandler_t
SIGFUNC is mainly used as a type handled by signal()
resp. sigaction(). This is all replaced by void (*)(int).

sighandler_t itself is a GNU extension which we do not use here for
better portability.

The only place where SIGFUNC is still there is one place zxwhen.c,
because here the call signature is different.