Rename ecl canonical name to irafcl
This is the name that is used in Debian and other Linux
distributions. Rationale is that the names cl and ecl are conflicting
with command names of other tools that may be installed on these
systems (Common Lisp interpreters). For compatibility, symbolic links
to cl and ecl are still provided.