Provide compatibility link to mkfloat.csh
Traditionally, the "mkfloat" script was a C shell, which was indicated
by its suffix. This was converted to a POSIX shell script long time
ago, without suffix.

The script however may be run during the build of external packages,
and they may refer to `mkfloat.csh`. To allow an easier migration, we
provide a compatiblity link here.

Although there is no known package that uses mkfloat.sh, we also
provide it here.