Use explicit pointer cast for zopnnd call in zfiobf.c
The first argument in ZOPNND() should really be declared as a `char *`.
However it is a SPP called function and therefore limited to the SPP
data types; so the first argument is declared as `PKG_CHAR *` (which
is a `short int`). Therefore we need to cast it when called from C.