Don't print some warnings in mkpkg's $ifolder/$ifnewer
Specifically, don't print warnings if the first file does not exist in
$ifolder, or of the second file does not exist in $ifnewer.

Reason is that $ifolder/$ifnewer are used to re-create a file when it
is older than its origin. Usually however the generated file does not
exist at all in the beginning; this is a normal case and shall not
cause a warning.