tbtables: append the overwritten extension rather than overwrite in-place.
In the original version the FITS file can become corrupted when
overwriting a bintable.  This wasn't fully diagnosed but appeared to
be related to the structure not properly tracking the offsets with the
actual FIO file position.  As a workaround, we will simply append the
overwritten extension rather than overwrite in-place.